requirements.txt
aioredis==1.0.0
asgiref==2.2.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
certifi==2018.1.18
channels==2.0.2
channels-redis==2.1.1
constantly==15.1.0
daphne==2.1.0
Django==2.0.3
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
msgpack==0.5.6
pytz==2018.3
six==1.11.0
Twisted==17.9.0
txaio==2.9.0
zope.interface==4.4.3
google==2.0.1
gcloud==0.18.3
google-auth==1.6.2
google-cloud-speech==0.36.3
grpc-google-cloud-speech-v1beta1==1.0.1
numpy==1.16.1
wave
